% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{heatmap}
\alias{heatmap}
\title{heatmap}
\format{\preformatted{'data.frame':	144 obs. of  5 variables:
 $ Var1 : Factor w/ 9 levels "a","b","c","d",..: 1 2 3 4 5 6 7 8 9 1 ...
 $ Var2 : Factor w/ 16 levels "A","B","C","D",..: 1 1 1 1 1 1 1 1 1 2 ...
 $ value: num  -0.622 0.271 0.563 2.426 3.788 ...
 $ X1   : Factor w/ 9 levels "b","a","d","c",..: 2 1 4 3 5 6 7 8 9 2 ...
 $ X2   : Factor w/ 16 levels "P","O","M","I",..: 13 13 13 13 13 13 13 13 13 14 ...
}}
\usage{
heatmap
}
\description{
A heatmap example
}
\keyword{datasets}

