% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesanderson.R
\docType{package}
\name{wesanderson}
\alias{wesanderson}
\alias{wesanderson-package}
\title{wesanderson}
\description{
list of palettes from Wes Anderson movies
}
\details{
list of palettes from Wes Anderson movies
}
