% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{wes_palettes}
\alias{wes_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 16.}
\usage{
wes_palettes
}
\description{
Use \code{\link{wes_palette}} to construct palettes of desired length.
}
\keyword{datasets}
