% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{wesanderson}
\alias{wesanderson}
\alias{wesanderson-package}
\title{wesanderson}
\description{
wesanderson
}
\details{
list of palettes from Wes Anderson movies, inspired by \href{http://wesandersonpalettes.tumblr.com/}{Wes Anderson palettes} which provides some useful palettes from Wes Anderson films.
}

